requirements.txt
dotenv==0.9.9
mysql-connector-python==9.2.0
pyaes==1.6.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.1.0


pyfiglet==1.0.2
asyncio==3.4.3
pyrofork==2.3.58
tgcrypto-pyrofork==1.2.7
configparser==6.0.1
PyYAML==6.0.2
pytz==2024.2